requirements.txt
# Basic libs
numpy==1.24.1
Pillow==9.5.0
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.1
scipy==1.13.1
einops==0.8.0
wandb==0.17.0  

# Diffusers
transformers==4.41.1
accelerate==0.30.1
diffusers==0.28.0

# Additional libs
Cython==3.0.10
cvxopt==1.3.2
pyopencl==2024.2.2
websockets==12.0
pylette==2.0.1